requirements.txt
discord.py
cython
setuptools
sqlalchemy[asyncio]

mypy
discord.py-stubs
